# apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WMS API",
    "version": "1.0.0",
    "description": "Minimal WMS API with Swagger UI"
  },
  "servers": [
    {
      "url": "http://localhost:3100"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/outbound-orders/{id}/boxes/{boxId}/items": {
      "get": {
        "summary": "List outbound box items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "boxId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Outbound box not found"
          }
        }
      },
      "put": {
        "summary": "Replace outbound box items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "boxId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "items"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "outbound_item_id",
                        "packed_qty"
                      ],
                      "properties": {
                        "outbound_item_id": {
                          "type": "integer"
                        },
                        "packed_qty": {
                          "type": "integer",
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Validation/box item error"
          },
          "404": {
            "description": "Outbound box not found"
          }
        }
      }
    },
...
    "/outbound-items": {
      "get": {
        "summary": "List outbound items",
        "parameters": [
          {
            "name": "outbound_order_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create outbound item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "outbound_order_id",
                  "product_id",
                  "qty"
                ],
                "properties": {
                  "outbound_order_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "lot_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "location_id": {
                    "type": "integer"
                  },
                  "qty": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "box_type": {
                    "type": "string"
                  },
                  "box_count": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          }
        }
      }
    },
    "/outbound-items/{id}": {
      "get": {
        "summary": "Get outbound item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update outbound item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "outbound_order_id",
                  "product_id",
                  "qty"
                ],
                "properties": {
                  "outbound_order_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "lot_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "location_id": {
                    "type": "integer"
                  },
                  "qty": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "box_type": {
                    "type": "string"
                  },
                  "box_count": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete outbound item (soft)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}